# Reduce Constraint object footprint for faster `clone` performance