--- a/src/solver/Constraint/GeneralCellPairConstraint.js
+++ b/src/solver/Constraint/GeneralCellPairConstraint.js
@@ -20,15 +20,15 @@ export class GeneralCellPairConstraint extends Constraint {
         let changed = false;
         for (let cellPair of this.cellPairs) {
             const [cell1, cell2] = cellPair;
-            const valueUsed1 = Array.from({ length: this.size + 1 }, () => false);
-            const valueUsed2 = Array.from({ length: this.size + 1 }, () => false);
-            for (let value1 = 1; value1 <= this.size; value1++) {
+            const valueUsed1 = Array.from({ length: board.size + 1 }, () => false);
+            const valueUsed2 = Array.from({ length: board.size + 1 }, () => false);
+            for (let value1 = 1; value1 <= board.size; value1++) {
                 if (!hasValue(board.cells[cell1], value1)) {
                     continue;
                 }
                 const cell1Candidate = board.candidateIndex(cell1, value1);
 
-                for (let value2 = 1; value2 <= this.size; value2++) {
+                for (let value2 = 1; value2 <= board.size; value2++) {
                     if (!hasValue(board.cells[cell2], value2)) {
                         continue;
                     }
@@ -86,15 +86,15 @@ export class GeneralCellPairConstraint extends Constraint {
                     continue;
                 }
 
-                const valueUsed1 = Array.from({ length: this.size + 1 }, () => false);
-                const valueUsed2 = Array.from({ length: this.size + 1 }, () => false);
-                for (let value1 = 1; value1 <= this.size; value1++) {
+                const valueUsed1 = Array.from({ length: board.size + 1 }, () => false);
+                const valueUsed2 = Array.from({ length: board.size + 1 }, () => false);
+                for (let value1 = 1; value1 <= board.size; value1++) {
                     if (!hasValue(board.cells[cell1], value1)) {
                         continue;
                     }
                     const cell1Candidate = board.candidateIndex(cell1, value1);
 
-                    for (let value2 = 1; value2 <= this.size; value2++) {
+                    for (let value2 = 1; value2 <= board.size; value2++) {
                         if (!hasValue(board.cells[cell2], value2)) {
                             continue;
                         }