Perf speedup on "PUZZLE NO. 395 : X SUMS - LITTLE KILLER HYBRID SUDOKU!": https://f-puzzles.com/?id=yt6qpjw8

210ms/"Check" --> 164ms/"Check"
(21030ms for 100 iterations --> 3277ms for 20 iterations)